BoardクラスからCellsクラスを分離
Boardクラスはゲームの状態を管理する
