+++ lib/loader.rb
@@ -2,6 +2,7 @@ require './lib/app.rb'
 require './lib/errors.rb'
 require './model/move.rb'
 require './model/board.rb'
+require './model/cells.rb'
 require './model/cell.rb'
 require './model/resource.rb'
 require './controller/request_handler.rb'
