+++ view/index.html.erb
@@ -20,7 +20,7 @@
       <tbody>
         <% (1..8).each do |y| %>
           <tr>
-            <% @board.row(y).each_value do |cell| %>
+            <% @board.cells.row(y).each_value do |cell| %>
               <td class="cell<%= ' movable' if @move.moves.has_key? cell.index %> cell_<%= cell.index %>" id="cell_<%= cell.index %>" data-index="<%= cell.index %>">
                 <span class="disc <%= cell.color.send(:to_s) %>"></span>
               </td>
